Allow setting priority for list entries
When we switched to using concat, our list entries started to be sorted
by name as opposed to just printed in the order specified in the hiera.
There is no way of working around it other than providing a way to
specify a higher priority manually in the hiera entry per list.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>